+++ moor/lib/moor.dart
@@ -12,7 +12,6 @@ export 'package:moor/src/runtime/query_builder/query_builder.dart';
 
 export 'package:moor/src/runtime/executor/connection_pool.dart';
 export 'package:moor/src/runtime/executor/executor.dart';
-export 'package:moor/src/runtime/executor/transactions.dart';
 export 'package:moor/src/runtime/custom_result_set.dart';
 export 'package:moor/src/runtime/data_verification.dart';
 export 'package:moor/src/runtime/data_class.dart';