+++ moor/lib/src/runtime/api/dao_base.dart
@@ -8,10 +8,7 @@ part of 'runtime_api.dart';
 /// For details on how to write a dao, see [UseDao].
 /// [T] should be the associated database class you wrote.
 abstract class DatabaseAccessor<T extends GeneratedDatabase>
-    extends DatabaseConnectionUser with QueryEngine {
-  @override
-  final bool topLevel = true;
-
+    extends DatabaseConnectionUser {
   /// The main database instance for this dao
   @override
   final T attachedDatabase;