+++ moor/lib/src/runtime/executor/transactions.dart
@@ -1,15 +1,11 @@
+import 'package:meta/meta.dart';
 import 'package:moor/moor.dart';
 import 'package:moor/src/runtime/executor/stream_queries.dart';
 
 /// Runs multiple statements transactionally.
-///
-/// Moor users should use [QueryEngine.transaction] to use this api.
-@Deprecated(
-  "This class will be private in moor 4. If you're using this class directly, "
-  'please describe your usage in https://github.com/simolus3/moor/issues/810.',
-)
-class Transaction extends DatabaseConnectionUser with QueryEngine {
-  final QueryEngine _parent;
+@internal
+class Transaction extends DatabaseConnectionUser {
+  final DatabaseConnectionUser _parent;
 
   @override
   GeneratedDatabase get attachedDatabase => _parent.attachedDatabase;
@@ -91,18 +87,15 @@ class _TransactionStreamStore extends StreamQueryStore {
 ///
 /// To use this api, moor users should use the [MigrationStrategy.beforeOpen]
 /// parameter inside the [GeneratedDatabase.migration] getter.
-@Deprecated(
-  "This class will be private in moor 4. If you're using this class directly, "
-  'please describe your usage in https://github.com/simolus3/moor/issues/810.',
-)
-class BeforeOpenRunner extends DatabaseConnectionUser with QueryEngine {
-  final QueryEngine _parent;
+@internal
+class BeforeOpenRunner extends DatabaseConnectionUser {
+  final DatabaseConnectionUser _parent;
 
   @override
   GeneratedDatabase get attachedDatabase => _parent.attachedDatabase;
 
-  /// Creates a [BeforeOpenRunner] from a [QueryEngine] and the special
-  /// [executor] running the queries.
+  /// Creates a [BeforeOpenRunner] from a [DatabaseConnectionUser] and the
+  /// special [executor] running the queries.
   BeforeOpenRunner(this._parent, QueryExecutor executor)
       : super.delegate(_parent, executor: executor);
 }