+++ moor/lib/src/runtime/query_builder/statements/delete.dart
@@ -3,8 +3,9 @@ part of '../query_builder.dart';
 /// A `DELETE` statement in sql
 class DeleteStatement<T extends Table, D extends DataClass> extends Query<T, D>
     with SingleTableQueryMixin<T, D> {
-  /// This constructor should be called by [QueryEngine.delete] for you.
-  DeleteStatement(QueryEngine database, TableInfo<T, D> table)
+  /// This constructor should be called by [DatabaseConnectionUser.delete] for
+  /// you.
+  DeleteStatement(DatabaseConnectionUser database, TableInfo<T, D> table)
       : super(database, table);
 
   @override